+++ OutOfHome.DataProviders.Boards.Grids.Downloader/Entities/Bigmedia/Common/InternalBmaBoard.cs
@@ -66,7 +66,7 @@ internal class InternalBmaBoard
     //public string Streets { get; set; }
 
     [JsonPropertyName("price")]
-    public float? Price { get; set; }
+    public int? Price { get; set; }
 
     [JsonPropertyName("supplier_sidetype")]
     public string SupplierSidetype { get; set; }
